requirements.txt
pillow
numpy
opencv-contrib-python
leveldb
google
grpcio
grpcio-tools